Fix(?) problem with shadowing by function param and addInst
I didn't actually find any problem with the existing impl, but it
seemed like it was buggy. Consider the case:

(define (id x) x)
(define foo (lambda id (id 1)))

Now when (id 1) is encountered and an instantiation is to be added for
id, what ought to happen is nothing, since id is a parameter, but
there hasn't actually been any kind of shadowing in the `insts` Map,
so the inst will be added to the global `id`, I think.