Fix sizeof in Codegen by generatin datatypes ahead of time
Should have been obvious -- of course we can't know the size of a type
if some field refers to an undefined datatype!

There is one suboptimality about this implementation. Overall, the
datatypes will be defined twice. Once in Codegen just for sizeof to
work, and once in withModuleFromAST call in Compile when actually
compiling the module.