Detect instantiations of datatypes in monotype, not monoCtion
monoCtion only detects when we construct types, which is no-good if
we, say, receive a datatype through FFI and send it back through
FFI. No construction or desctruction locally. By detecting insts in
monotype, we cover all cases.

Also, it seems I accidentally comitted the crucial monotype part a
couple of commits ago. Sorry for jumbling the history xo