Don't load to local from stack and vice versa until needed
Sort of a "lazy" way of stack-allocating/loading variables that
decreases the number of instructions generated by a some
percents. Eliminates a pattern I've observed in generated code where
variables are loaded to locals, only to be immediately put on the
stack again to be sent as reference params or some such.