Remove character literals. Less confusion for users, I think
There is no one correct definition of "character". Valid definitions
include: ASCII character, Unicode codepoint, Unicode grapheme
cluster.

UTF-32 codepoints make quite a bit of sense -- it's fixed width and
can represent almost all common "symbols", unlike ASCII which is very
US-centric. The problem is that some things that render as a single
glyph to the user are actually composed of multiple codepoints. These
are "grapheme clusters". For example, some emojis are actually
grapheme clusters rendered as a single glyph. The codepoints `brown
skin tone` and `smiling face` in sequence often render as a single,
brown-skinned smiling face.

Different existing languages handle characters in different ways as
well.

I think we'll just stick with strings for now, and not recognizing
that there's a single-correct "element" type of strings, so to
say. There should be functions to convert strings sequences of bytes, of
codepoints, and of grapheme clusters, but no single "iterate"
instance, for example.