Add `carth version`. Bake package.yaml version & commit into binary
Bake uses templatehaskell. At compiletime of the carth compiler,
version numbers are read from package.yaml and commit hash & date is
gotten from shell command.

Something to keep in mind is that the file with the baked values is
not recompiled by default. Luckily it was solved easily. Any
modification of package.yaml forces rebuild, so changing version in
the file will change version in the binary. To force recompilation
when git status changes, use templatehaskell "command"
`qAddDependentFile ".git/index"`, which seems to add that file to some
watchlist in the metadata of the compiled library? It works anywho.