Got OrcJIT to work! Look in process address space in resolver.
Previously, I was trying to resolve everything with
CompileLayer.findSymbol, however, this function only looks for symbols
in the compile-layer (i.e. LLVM module). LingingLayer.findSymbol
also includes manually added .o files, but this is also not enough. To
get parity with the MCJIT version, we have to look in the address
space of the running process. This has to be done with
`getSymbolAddressInProcess` (after `loadLibraryPermanently Nothing`).