Fix bug where recursive fun in let didn't populate captures properly
I hadn't before taken into account that the closure has to be defined
before the captures structure could be populated when the function is
recursive.

Now `genLet` has basically been split into two passes, where the first
pass defines variables and functions on the stack, but doesn't
populate closure captures, instead just heap allocating an empty
captures struct for each closure and populating the capture structs in
the second pass when all closures are defined and available in the
environment.

`genLet` was actually did two passes before as well, but that was to
handle non-topological order, and not recursion. This isn't an issue
anymore since a few commits ago though, so it was basically removed
now.