Update stackage release to lts-15.1
Had some trouble building the thing on new laptop, both on Guix and
with Stack on Arch. I couldn't find much info about it, but one person
suggested memory error. I hoped this wasn't the case, since the laptop
is brand new, but that it instead was some kind of GHC/Cabal
bug. Indeed, the GHC version in both cases was 8.6.something. Updating
to 8.8.something with LTS 15.1 seems to have helped! While the
miscompilation occured a bit randomly before, it was still likely to
happen, but now the compilation succeeded on the first try, so that
was probably it!