Fix externs having fastcc & return differently when in tail pos
This is part of improving guarantees of tail call optimization. Still
a little bit to go.

I think the only remaining thing is to change {} to void, as it seems
LLVM doesn't like tail calls otherwise. If we return an int, it's all
good, but if we return unit, we get stack overflow.