Return void instead of {}. Allows for more tail call optimization
If a function returns {} (i.e. Unit), then

    %x = call {} @foo()
    ret {} %x

is automatically converted by LLVM to something like

   call {} @foo()
   ret {} zeroinitializer

This is a problem, as tail calls are only optimized if ["ret
immediately follows call and ret uses value of call or is
void"](http://llvm.org/docs/CodeGenerator.html#tail-call-optimization). Since
the value (%x) is no longer used after the conversion (zeroinitializer
instead), and {} is not void, the optimization doesn't happen. I
observe this in practice.

So to fix this issue, we either have to get LLVM to not do that stupid
conversion, or return void instead. One way to get LLVM not to do the
inversion would be to represent Unit as e.g. an i8. Then LLVM will be
forced to use %x, as it doesn't know that it's always zero. This
method would likely work, as I have observed that tail calls are
indeed optimized for functions returning integers. However, this would
make units sized, which seems like it would cause trouble or at least
prevent some optimizations.

So I went with the alternative: return void when the type is Unit or
any other single-variant enum. Luckily, not much code needed to be
changed (essentially, use genRetType instead of genType where
appropriate), but the solution is not differently typed, so I'll have
to be a bit careful.