Split let into let1, let, letrec
let1: Bind one variable to one value in body. Can shadow.

    (let1 x 3
      (* x x))

let: Sugar for nester let1:s, which means it's not like the old let,
in that it shadows, requires bindings to be in order, and cannot be
recursive.

    (let1 x 3
      (let ((x x)
            (y x)
            (x 5)
            (z x))
        (cons y z)))
    ;; Ok: (Cons 3 5)

letrec: Pretty much exactly like the old let. Bindings exist in the
scope of their own bodies, so functions can be recursive and shadowing
is not possible. If a non-function variable is found to be recursive,
the type checker throws an error. Bindings are automatically ordered
in topological order.

    (let1 x 3 (letrec ((x x)) x))
    ;; Error

    (letrec ((x y)
             (y 3)
             (fact (fmatch (case 0 1) (case n (* n (fact (dec n))))))
             (foo (fun (x) (bar (dec x))))
             (bar (fmatch (case 0 1337) (case x (foo x)))))
      (cons* x (fact 5) (foo 100)))
    ;; Ok: (Cons* 3 120 1337