TNat8, TNat16, ... -> (TNat 8), (TNat 16), ... & add TNatSize
Instead of having hard-coded sized variants for integer types, just
have a bit-width field. Fewer variants => better, also this will allow
for arbitrary bit-width integers if we want that (I think it could be
nice).

Also add the T{Nat,Int}Size variants, which look like {Nat,Int} in
carth source. They are like {u,i}size in Rust and size_t in C -- an
integer of the same size as a pointer. 32 bits on 32-bit platforms,
and 64 bits on 64-bit platforms. Won't actually work for 32-bit right
now, as a bunch of stuff is still hard-coded to 64, but that can be
changed in the future.