Fix partial app of builtin virtual for non-literal args
The values of the args were generated in the local scope, but the
application happened in the scope of the generated lambda. The
difference of scope made the local references to the arg vals
invalid. See the TODO for more details -- the current implementation
of genAppBuiltinVirtual is really quite wasteful and generates
redundant functions.