Fix lambda body free var capturing being too greedy
lambdaBodyFreeVars was both overcomplicated and wrong. Instead of
excluding globals that are not overridden by locals, just do an
intersection with locals -- the whole point is that we have to
capture *local* free variables, but we don't have to do that with
globals because they are **global**!! The bug was that builtins like
`+`, `transmute`, etc, were not excluded, so they were captured in
their curried states, even if the calls to them later were actually
saturated! This meant needless overhead.