Fix builtin virtual calls being lambda wrapped even when saturated
So e.g. `+` was applied like ((+-wrapper a) b) instead of
like (+-operator [a b]), kind of.

This fix together with the previous commit reduces the amount of
closure-juggling logic going on, and improves performance quite a bit
in many cases. As an example, the runtime for (ackermann 4 1) went
from something like 260s to around 50s. Still a lot of overhead
though, compared to the C version which took like 2s...