Repr Unit as [0xi8]. Alt to dceded5 for fixing TCO for Unit ret funs
This method feels less edge-case-y. Effect is the same.

As a refresher, the problem was that for tail recursive functions with
return type Unit, LLVM would optimize
  %x = tail call {} foo()
  ret {} %x
to
  %x = tail call {} foo()
  ret {} zeroinitializer
However, this "optimization" would prevent a later tail-call
optimization from happening, as TCO requires us to return either void,
or the local of the return value of the last call. zeroinitializer may
have the same value as %x, but the optimizer doesn't recognize it as
fulfilling the conditions, so TCO didn't happen. By representing Unit
as an empty array instead, this problem doesn't happen, because %x is
simply not replaced with zeroinitializer when the type is an array
type. Don't know why this is the case, but it works out.