Fix workaround for JIT not detect. glob vars by calling GC_add_roots
Boehm GC is smart, and normally detects global variables as roots by
scanning the static segment of the ELF executable. However, when
JITtin in LLVM, there is no such segment, and all roots are not
detected. The previous workaround was to treat global non-function
vars basically like locals, and capture them in closures etc. This
guaranteed that they were always available on the stack, in a
register, or on the heap via another root. This was wasteful though,
and many closures that could've had empty envs had them filled with
crud.

New method simply calls GC_add_roots after initializing each global
var in init.