(mostly) implement checking of same size for transmute in Infer
Preferably, all kinds of checking should happen as early as possible,
so we don't have to remember tons of polluting SrcPos:s in the AST for
later stages. This is a big step in removing positioning from Codegen.

Also, this lays much of the groundwork for supporting typeclasses more
generally!

The implementation is not quite correct/done though. Indeed, it
rejects transmutes for a pair of monotypes not of the same size, but
the inferring and scheme creating is a little broken. See
~test/tests/bad/transmute-size-mismatch2.carth~.