Output fewer instructions
Does a few things.

1. Don't generate dummy closures for global functions. If we want to
call a global func, we can now do it directly, without loading and
extracting! The con, if it's even a con, is that now if we want to
pass the global func as an arg, we'll have to generate the dummy
closure wrapper on the fly. HOWEVER, this can be done in a constant
instruction, so likely still more efficient than the load from the
global reference we would previously still need to have done.

2. Don't load, store, and more for zero-sized types. Just don't output
any instruction, and return a ConstantOperand zero/undef/whatever value.

3. In `genStruct`, instead of using runtime `insertvalue` on every
single operand, first check if any of them are constant. If so, use a
constant `litStruct` with some/all pre-filled values for the initial
value to append to instead.

4. jesus im tired. Wakiing up tomorrow won't be easy...

4 fr tho. `genIndexStruct` on zero/undef vals will just return a
constant zero/undef val of the right type instead of generating
gep/extractvalue instructions. On a `Struct` value, extract the value
now at comptime. Consutanto rabu!