Fix segfault when representing variant contained implicit padding
Before the fix, a type like `(Either (Fun _ _) [Int8 Int])` would be
repersented by the `[Int8 Int]` variant. The 8 bytes of the captures
pointer in the Fun variant would occupy the same space as the Int8 + 7
byte padding of the representing variant. When passed as an argument
to a function, LLVM would think that the padding was unused space,
since it was implicit, so it would disregard it make it zeroed when
passing the Int8 (I assume through a register). This would of course
destroy the captures pointer, if the value was really of that variant.

This commit fixes the bug by changing how tagged unions are
represented in LLVM. Now, a datatype with more than one variant (and
them not nonempty) is represented in LLVM by a representing struct of
3 elements. First, the tag. Second, a byte-array to achieve the
correct size and ensure all space is explicitly seen as used by
LLVM. Third, an empty array of aligning integer type, to ensure that
the alignment of the representing type is as big as the greatestly
aligned variant.