Update std lib with io & networking stuff and add gemini client ex.
Also some other updates to std, as prerequisites for the gemini
implementation. The gemini example uses some simple networking code
and works as a sort of regression test for the recent struct layout
fixes, since it was when I started working on this that FFI stuff
broke way back when xd.