repr closure with named, genericized type "closure" & refactoring
I realized, in C we can't do anonymous structs the same way we can in
LLVM.

Sure, this is a legal statement

    struct { int a; } foo = { 1337 };

But following it with this is not

    struct { int a; } bar = foo;

This is because C treats each anonymous struct as a different,
incompatible type, even if they are structurally equal. This won't
work for us, so instead of representing a closure like

    struct { void* captures; Bar (*fun)(void*, Foo); } f = ...;

we'll do

    closure f = ...;

where

    typedef struct { void* captures; void* fun; } closure;

This will only incur a "cost" at compile time, with us needing to
insert a bitcast to the actual function type before calling it. And
while the new LLVM will be slightly more opaque, it should also be
significantly less verbose.