Fix matchPat accepting over-saturated paren list
Bug stems from zipWithM ignoring the remainder of the longer list when
uneven. Encountered this in the `do` macro, where in

    (do ...
        (let1 lhs rhs body)
        foo)

`body` was ignored, and instead of being interpreted as an expression,
the let1-form was interpreted as the let1-case in in the do-macro. So
`body` was discarded. I noticed this when both `body` and `foo` were
applications of `display`, but only the latter line was actually
displayed.