std: give parsers to use slightly more descriptive error messages
Not just None now, but a Left containing the remaining input and a
simple message. From the length of the remaining input, we can deduce
location in source. Line, column, etc.

Next improvement would be to gather a set of things we could have seen
here. So we get something like "expected '}', identifier, or integer".