lower: Make type env part of State. Want to define structs OTG
Since anonymous structs would be a no-go in a potential C backend, we
want to generate named definitions for each pseudo-anonymous struct we
ever need, like those for closure captures. For this to work, the type
environment must be appendable and thus a part of the State of
Lower. This has some monad juggling implications on existing code.

Also, begin working on lowering closures.