Fix loop when lowering recursive definitions
We did it this way before the lowering rewrite as well. I guess I was
just hoping that there might not be any loop issue this time around,
since some things are a bit different. I didn't actually bother
thinking it through though, so that we still needed to separate
declaration and definition of captures this way is not actually
surprising.