Fix bugs due to inconsistencies in keeping things on stack / in reg
All structs are kept on the stack for the most part, except for right
when we need to store them in another struct or pass them as arguments
to a function. Then we might need to load them, in the case of a
function call, depending on whether the struct is big enough to
require a pass-by-reference. This is all a bit tricky to keep track
of, but I hope I've fixed all bugs with these changes, and that the
implementation is clear *enough*.