Use frontend sizeof when defining datatypes in Lower
Basically, I got stuck with recursion bugs when trying to define
datatypes and the programs types in the same "step". With these
changes, defining datatypes, which includes looking at whether types
are sized and lowering them to Low.Type, now only uses
Monomorphic.Type with Sizeof.sizeof for computing sizes. The <<loop>>
bugs are gone!

Maybe datatype definition should be in a separate module? It is a
little noisy to have multiple, slightly different versions of
e.g. `lowerType` and `sizeof` in the same module, albeit in different
scopes.

Also, I unified the representation of types in the frontend while I
was at it, so the same `Sizeof.sizeof` now works on types from Parsed,
Inferred, Checked, and Monomorphic!