Show warning when declaring extern with non-C compatible identifier
The kebab-case identifiers of Carth cannot be spelled out in C without
mangling, and we don't want to mangle `extern`s -- that would be very
surprising behaviour! Thus, we now print warnings when declaring an
`extern` with any characters not valid in a C identifier, e.g. the
dashes (-) of kebab-case.

It is still legal to use non-C compatible identifiers though, since
one may be linking with e.g. a LLVM or Rust library instead, where
more colorful identifiers are permitted. In Rust we can for example
export a function with kebab-case using `#[export_name("show-int")]`.