All alternative, external test runner: run-tests.sh
This is a very simple bash script runs some of the essential
end-to-end tests. It's an alternative to test/SystemSpec.hs that
doesn't have any issues with libgc failing with "collecting from
unknown thread" or whatever it was exactly. This versin of the runner
can leave the GC on when compiling, which allows for longer running
tests.