Mangle names of globals / functions
To play better with the C backend mostly. I had the issue that my own
`abs` shadowed / redeclared the builtin `abs`, and with a different
type signature as well.

Now, all globals are prefixed with "_C<LEN>H". Identifiers beginning
with underscore followed by uppercase letter is reserved in C, so that
means there ought not to be any conflicts with user written functions
when we link against external symbols.