Pass fun args on stack correctly & always pass loop args by val
It turns out that big structs in C aren't actually (necessarily?)
passed by reference, but rather passed on the stack specifically. It
may thus be incorrect to treat what I called a "pass by reference"
argument just the same as a pointer value argument, and give it a heap
pointer. I checked it out on godbolt, and unless my mind tricked me,
that was the situation.

The other thing is that Loop arguments must be copies on each
iteration, just like a stack frame has to be copied in
recursion. Otherwise we reuse allocations in a bad way.