std: make array/iter safe by panicing on out-of-bounds
Note that it shouldn't ever be able to even reach this, unless there's
a compiler bug or a bug in xrange in stdlib, but the fact is that
there well may be! This way, at least I get a slightly useful error
message from my tests when I've goofed up something in the compiler.