Begin work on new selfhosted, query-based compiler
I have no idea how far I'll get, or how painful it will be with Carth
in the state that it is, but I figure there's no better way to
identify the pain points than to experience the pain directly!

Also, I'm hoping that the on-demand architecture and a
reimplementation of the type system to use bidirectional type checking
will make adding type classes, modules, & comptime evaluation much
easier.